+++ src/store/desktop_oauth.rs
@@ -94,10 +94,12 @@ pub(crate) async fn insert_device_code(
     ttl_seconds: i64,
 ) -> AuthStackResult<()> {
     initialize_schema_async().await?;
+    // `::bigint` so Spin's JSON→Int64 bind matches the prepared parameter.
+    // A bare INTEGER column (or `::int4`) fails with WrongType Int4 vs i64.
     execute_sql(
         "INSERT INTO buwiz_server.oauth_device_codes (\
             device_code_hash, user_code, client_id, verification_uri, interval_seconds, expires_at \
-         ) VALUES (?1, ?2, ?3, ?4, ?5, to_timestamp(?6))",
+         ) VALUES (?1, ?2, ?3, ?4, ?5::bigint, to_timestamp(?6))",
         vec![
             json!(device_code_hash),
             json!(user_code),
